move NodeContent to presenter package
