+++ Utils/azuretools-core/src/com/microsoft/azuretools/core/model/AzureMvpModelHelper.java
@@ -22,17 +22,15 @@
 
 package com.microsoft.azuretools.core.model;
 
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Set;
+
 import com.microsoft.azure.management.Azure;
-import com.microsoft.azure.management.redis.RedisCache;
 import com.microsoft.azure.management.redis.RedisCaches;
 import com.microsoft.azuretools.authmanage.AuthMethodManager;
 import com.microsoft.azuretools.authmanage.SubscriptionManager;
 import com.microsoft.azuretools.sdkmanage.AzureManager;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Set;
 
 public class AzureMvpModelHelper {
     
@@ -51,8 +49,8 @@ public class AzureMvpModelHelper {
      * @return A map containing RedisCaches with subscription id as the key
      * @throws IOException getAzureManager Exception
      */
-    public HashMap<String, ArrayList<NodeContent>> getRedisCaches() throws IOException {
-        HashMap<String, ArrayList<NodeContent>> redisCacheMaps = new HashMap<String, ArrayList<NodeContent>>();
+    public HashMap<String, RedisCaches> getRedisCaches() throws IOException {
+        HashMap<String, RedisCaches> redisCacheMaps = new HashMap<String, RedisCaches>();
         AzureManager azureManager = AuthMethodManager.getInstance().getAzureManager();
         if (azureManager == null) {
             return redisCacheMaps;
@@ -67,12 +65,7 @@ public class AzureMvpModelHelper {
             if (azure == null || azure.redisCaches() == null) {
                 continue;
             }
-            ArrayList<NodeContent> nodeContentList = new ArrayList<NodeContent>();
-            for (RedisCache redis: azure.redisCaches().list()) {
-                NodeContent nodeContent = new NodeContent(redis.id(), redis.name(), redis.provisioningState());
-                nodeContentList.add(nodeContent);
-            }
-            redisCacheMaps.put(sid, nodeContentList);
+            redisCacheMaps.put(sid, azure.redisCaches());
         }
         return redisCacheMaps;
     }
