+++ Utils/azuretools-core/test/com/microsoft/azuretools/core/model/AzureMvpModelHelperTest.java
@@ -22,36 +22,33 @@
 
 package com.microsoft.azuretools.core.model;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Set;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
+import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
-import com.microsoft.azure.Page;
-import com.microsoft.azure.PagedList;
 import com.microsoft.azure.management.Azure;
-import com.microsoft.azure.management.redis.RedisCache;
 import com.microsoft.azure.management.redis.RedisCaches;
 import com.microsoft.azuretools.authmanage.AuthMethodManager;
 import com.microsoft.azuretools.authmanage.SubscriptionManager;
 import com.microsoft.azuretools.sdkmanage.AzureManager;
-import com.microsoft.rest.RestException;
-
-import org.powermock.api.mockito.PowerMockito;
-
-import static org.junit.Assert.*;
-import static org.mockito.Mockito.*;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 
 @RunWith(PowerMockRunner.class)
@@ -82,31 +79,13 @@ public class AzureMvpModelHelperTest {
     @Mock
     private SubscriptionManager subscriptionManagerMock;
     
-    private class MockedPagedList<RedisCache> extends PagedList<RedisCache> {
-
-        @Override
-        public Page<RedisCache> nextPage(String nextPageLink) throws RestException, IOException {
-            return null;
-        }
-    }
-    
     private static final String MOCK_SUBSCRIPTION = "00000000-0000-0000-0000-000000000000";
     private static final String MOCK_REDIS_ID = "test-id";
     
-    
     @Before
     public void setUp() throws IOException {
-
-        Set<String> mockSidList = Stream.of(MOCK_SUBSCRIPTION).collect(Collectors.toSet());
-        
         PowerMockito.mockStatic(AuthMethodManager.class);
         when(AuthMethodManager.getInstance()).thenReturn(authMethodManagerMock);
-        when(authMethodManagerMock.getAzureManager()).thenReturn(azureManagerMock);
-        when(azureManagerMock.getAzure(anyString())).thenReturn(azureMock);
-        when(azureManagerMock.getSubscriptionManager()).thenReturn(subscriptionManagerMock);
-        when(subscriptionManagerMock.getAccountSidList()).thenReturn(mockSidList);
-        when(azureMock.redisCaches()).thenReturn(redisCachesMock);
-        when(redisCachesMock.list()).thenReturn(new MockedPagedList<RedisCache>());
 
         azureMvpModelHelper = AzureMvpModelHelper.getInstance();
     }
@@ -124,8 +103,16 @@ public class AzureMvpModelHelperTest {
     @Test
     public void testGetRedisCaches() throws IOException {
         final int expectKeySetSize = 1;
+        final Set<String> mockSidList = Stream.of(MOCK_SUBSCRIPTION).collect(Collectors.toSet());
         
-        HashMap<String, ArrayList<NodeContent>> redisCachesMap = azureMvpModelHelper.getRedisCaches();
+        when(authMethodManagerMock.getAzureManager()).thenReturn(azureManagerMock);
+        when(azureManagerMock.getAzure(anyString())).thenReturn(azureMock);
+        when(azureManagerMock.getSubscriptionManager()).thenReturn(subscriptionManagerMock);
+        when(subscriptionManagerMock.getAccountSidList()).thenReturn(mockSidList);
+        when(azureMock.redisCaches()).thenReturn(redisCachesMock);
+        
+        HashMap<String, RedisCaches> redisCachesMap = azureMvpModelHelper.getRedisCaches();
+        
         assertNotNull(redisCachesMap);
         assertEquals(expectKeySetSize, redisCachesMap.keySet().size());
         assertNotNull(redisCachesMap.get(MOCK_SUBSCRIPTION));
@@ -134,8 +121,10 @@ public class AzureMvpModelHelperTest {
     @Test
     public void testGetRedisCachesWhenAzureManagerIsNull() throws IOException {
         final int expectKeySetSize = 0;
+        
         when(authMethodManagerMock.getAzureManager()).thenReturn(null);
-        HashMap<String, ArrayList<NodeContent>> redisCachesMap = azureMvpModelHelper.getRedisCaches();
+        
+        HashMap<String, RedisCaches> redisCachesMap = azureMvpModelHelper.getRedisCaches();
         assertNotNull(redisCachesMap);
         assertEquals(expectKeySetSize, redisCachesMap.keySet().size());
     }
@@ -143,9 +132,11 @@ public class AzureMvpModelHelperTest {
     @Test
     public void testGetRedisCachesWhenSubscriptionManagerIsNull() throws IOException {
         final int expectKeySetSize = 0;
+        
         when(authMethodManagerMock.getAzureManager()).thenReturn(azureManagerMock);
         when(azureManagerMock.getSubscriptionManager()).thenReturn(null);
-        HashMap<String, ArrayList<NodeContent>> redisCachesMap = azureMvpModelHelper.getRedisCaches();
+        
+        HashMap<String, RedisCaches> redisCachesMap = azureMvpModelHelper.getRedisCaches();
         assertNotNull(redisCachesMap);
         assertEquals(expectKeySetSize, redisCachesMap.keySet().size());
     }
@@ -153,10 +144,12 @@ public class AzureMvpModelHelperTest {
     @Test
     public void testGetRedisCachesWhenAzureIsNull() throws IOException {
         final int expectKeySetSize = 0;
+        
         when(authMethodManagerMock.getAzureManager()).thenReturn(azureManagerMock);
         when(azureManagerMock.getSubscriptionManager()).thenReturn(subscriptionManagerMock);
         when(azureManagerMock.getAzure(anyString())).thenReturn(null);
-        HashMap<String, ArrayList<NodeContent>> redisCachesMap = azureMvpModelHelper.getRedisCaches();
+        
+        HashMap<String, RedisCaches> redisCachesMap = azureMvpModelHelper.getRedisCaches();
         assertNotNull(redisCachesMap);
         assertEquals(expectKeySetSize, redisCachesMap.keySet().size());
     }
@@ -166,6 +159,7 @@ public class AzureMvpModelHelperTest {
         when(authMethodManagerMock.getAzureManager()).thenReturn(azureManagerMock);
         when(azureManagerMock.getAzure(anyString())).thenReturn(azureMock);
         when(azureMock.redisCaches()).thenReturn(redisCachesMock);
+        
         azureMvpModelHelper.deleteRedisCache(MOCK_SUBSCRIPTION, MOCK_REDIS_ID);
         verify(redisCachesMock, times(1)).deleteById(MOCK_REDIS_ID);
     }
@@ -173,6 +167,7 @@ public class AzureMvpModelHelperTest {
     @Test
     public void testDeleteRedisCacheWhenAzureManagerIsNull() throws IOException {
         when(authMethodManagerMock.getAzureManager()).thenReturn(null);
+        
         azureMvpModelHelper.deleteRedisCache(MOCK_SUBSCRIPTION, MOCK_REDIS_ID);
         verify(redisCachesMock, times(0)).deleteById(MOCK_REDIS_ID);
     }
@@ -181,6 +176,7 @@ public class AzureMvpModelHelperTest {
     public void testDeleteRedisCacheWhenAzureIsNull() throws IOException {
         when(authMethodManagerMock.getAzureManager()).thenReturn(azureManagerMock);
         when(azureManagerMock.getAzure(anyString())).thenReturn(null);
+        
         azureMvpModelHelper.deleteRedisCache(MOCK_SUBSCRIPTION, MOCK_REDIS_ID);
         verify(redisCachesMock, times(0)).deleteById(MOCK_REDIS_ID);
     }
@@ -190,6 +186,7 @@ public class AzureMvpModelHelperTest {
         when(authMethodManagerMock.getAzureManager()).thenReturn(azureManagerMock);
         when(azureManagerMock.getAzure(anyString())).thenReturn(azureMock);
         when(azureMock.redisCaches()).thenReturn(null);
+        
         azureMvpModelHelper.deleteRedisCache(MOCK_SUBSCRIPTION, MOCK_REDIS_ID);
         verify(redisCachesMock, times(0)).deleteById(MOCK_REDIS_ID);
     }
